file-conversion/markdown-to-pdf-agent/requirements.txt
markdown-it-py
fpdf2
python-frontmatter
rich
questionary
langchain
langchain-openai
python-dotenv
pytest
